Struct samd_dma::Channel

pub struct Channel { /* fields omitted */ }

DMA channel.

This structure represents a DMA channel. Using the start_transfer method, you can setup a DMA transfer transaction. You can also configure the DMA channel properties.

Methods

impl Channel

pub fn id(&self) -> u8

Return the channel ID.

pub fn set_run_standby(&mut self, run_standby: bool)

Configure whether the channel continues to run in standby.

pub fn get_run_standby(&self) -> bool

Return true if the channel continues to run in standby.

pub fn set_burst_length(&mut self, burst_len: BurstLength)

Configure how many beats are in a burst.

pub fn get_burst_length(&self) -> BurstLength

Get the length of a burst in beats.

pub fn set_trigger_action(&mut self, trig_act: TriggerAction)

Set the trigger action for the channel.

pub fn get_trigger_action(&self) -> TriggerAction

Get the trigger action for the channel.

pub fn set_fifo_threshold(&mut self, fifo_threshold: FifoThreshold)

Set threshold for when destination writes occur.

pub fn get_fifi_threshold(&self) -> FifoThreshold

Get the threshold for when destination writes will occur.

pub fn set_source(&mut self, source: TriggerSource)

Set the source trigger for the DMA Channel.

pub fn get_source(&self) -> TriggerSource

Get the trigger source for the channel.

pub fn set_priority(&mut self, priority: Priority)

Set the priority level of the channel.

pub fn get_priority(&self) -> Priority

Get channel priority level.

pub fn get_first_descriptor(&self) -> &mut TransferDescriptor

Get a mutable reference to the first descriptor for the channel.

pub fn get_interrupt_flags(&self) -> Interrupts

Get the channel's interrupt flags.

pub fn clear_interrupt_flags(&mut self)

Reset the channel's interrupt flags.

pub fn enable_interrupts(&mut self, interrupts: Interrupts)

Enable interrupts for the channel. Any interrupts that are not set will be disabled.

pub fn get_enabled_interrupts(&self) -> Interrupts

Get the set of enabled channel interrupts.

pub fn get_writeback_descriptor(&mut self) -> *mut TransferDescriptor

Read descriptor from the Write-back Address of this channel.

Safety

The write-back address of a DMA channel is volatile, as the DMA engine can change it at any time. As such, reading and writing to this descriptor is unsafe. Act cautiously.

pub fn enable(&mut self)

Enable the DMA channel.

After this call, this channel will be a part of the DMA arbitration scheme (if its corresponding priority level is active), and trigger events will cause the transaction to start from the first descriptor.

pub fn is_enabled(&self) -> bool

Return whether the channel is enabled or not.

pub fn reset(&mut self)

Reset the DMA channel. This will set all channel registers to their reset values.

If the channel is still enabled (or is in the process of being disabled), this command will be ignored.

pub fn trigger(&mut self)

Manually trigger the channel.

Interrupt Safety

This method is not interrupt-safe, and could lead to lost updates.

It is the responsibility of the caller to ensure that the call-site is in an interrupt-free section.

pub fn suspend(&mut self) -> bool

Suspend the ongoing transaction. Returns true if the command was successfully written, false if another command is ongoing.

This call returns immediately, but the suspend operation won't complete until the ongoing burst transfer completes.

pub fn resume(&mut self) -> bool

Resume the ongoing transaction. Returns true if command was successfully written, false if another command is ongoing.

pub fn disable(&mut self)

Disable the channel. This aborts any ongoing transaction.

This call returns immediately, but the transaction will not be aborted until the ongoing burst transfer completes.

pub fn is_pending(&self) -> bool

Returns true if a transfer is pending on the channel. Returns false if a channel trigger action is completed, a bus error is detected, or the channel is disabled.

pub fn is_busy(&self) -> bool

Returns true if the channel has started a transfer. Returns false if a channel trigger action is started, a bus error is detected, or the channel is disabled.

pub fn poll_status(&mut self) -> Result<WaitResult, TransactionError>

Poll the channel to determine the status of the transaction.

This will read and reset the interrupt flag registers of the channel to determine the status of the channel. Therefore it is expected to be called in the appropriate interrupt vector.

Any non-error state will return Ok(WaitResult). Any errors will be returned as Err(TransactionError).